move creating new user service, changing password service to user.js file
