+++ backend/controllers/user.js
@@ -1,3 +1,6 @@
+const mongoose = require("mongoose");
+const userManagementService = require("../services/userManagement.service");
+
 async function changeUserInfo(req, res) {
     const { newPhone, newEmail } = req.body;
     const userId = res.locals.userId;
@@ -18,3 +21,28 @@ async function changeUserInfo(req, res) {
         res.status(500).json({ message: error.message });
     }
 }
+
+const createUser = async (req, res) => {
+	try {
+		const userData = req.body;
+		const newUser = await userManagementService.createUser(userData);
+		return res.status(201).json(newUser);
+	} catch (error) {
+		res.status(error.status || 500).json({ message: error.message });
+	}
+};
+
+const changePassword = async (req, res) => {
+	try {
+		const data = req.body;
+		const result = await userManagementService.changePassword(data);
+		return res.status(200).json({ status: "076", result });
+	} catch (error) {
+		res.status(error.status || 500).json({ message: error.message });
+	}
+};
+
+module.exports = {
+	createUser,
+	changePassword
+};
